Refactor DataframeReader
