+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/ddl.scala
@@ -24,9 +24,11 @@ import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.catalog.CatalogTable
 import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
+import org.apache.spark.sql.catalyst.util.CaseInsensitiveMap
 import org.apache.spark.sql.errors.QueryCompilationErrors
 import org.apache.spark.sql.execution.command.{DDLUtils, LeafRunnableCommand}
 import org.apache.spark.sql.execution.command.ViewHelper.createTemporaryViewRelation
+import org.apache.spark.sql.execution.datasources.v2.DataSourceV2Utils
 import org.apache.spark.sql.internal.StaticSQLConf
 import org.apache.spark.sql.types._
 
@@ -88,9 +90,20 @@ case class CreateTempViewUsing(
     }
 
     val catalog = sparkSession.sessionState.catalog
-    val dfReader = sparkSession.read.format(provider).options(options)
-    val analyzedPlan = userSpecifiedSchema.map(dfReader.schema(_))
-      .getOrElse(dfReader).load().logicalPlan
+    val analyzedPlan = DataSource.lookupDataSourceV2(provider, sparkSession.sessionState.conf)
+      .map { tblProvider =>
+        DataSourceV2Utils.loadV2Source(sparkSession, tblProvider, userSpecifiedSchema,
+          CaseInsensitiveMap(options), provider)
+      }.flatten.getOrElse {
+        val dataSource = DataSource(
+          sparkSession,
+          userSpecifiedSchema = userSpecifiedSchema,
+          className = provider,
+          options = options)
+
+        Dataset.ofRows(
+          sparkSession, LogicalRelation(dataSource.resolveRelation()))
+      }.logicalPlan
 
     if (global) {
       val db = sparkSession.sessionState.conf.getConf(StaticSQLConf.GLOBAL_TEMP_DATABASE)
